// lambda/sm_autoplayer/src-ts/lib/move-enum.ts
export enum MoveScoresEnum {
    FROM_PILE = 1000,
    ADD_TO_CENTER_STACK = 100,
    OPEN_A_SPACE = 20,
    PLAY_FROM_HAND = 5,
    PLAY_FROM_STACK = 4,
    PLAY_JOKER=-2,
    DISCARD_IN_SEQUENCE = 13,
    DISCARD_BLOCK_SELF = -10,
    DISCARD_OUT_OF_SEQUENCE = 0,
    DISCARD_SAME_CARD = 5,
    DISCARD_BLOCK_SEQUENCE = -4,
    DISCARD_JOKER = -100,
    DISCARD_ON_JOKER = -200,
    REFRESH_HAND =500
}